+++ setup.cfg
@@ -123,7 +123,7 @@ ignore_errors = False
 [mypy-bokeh.util.sampledata]
 ignore_errors = False
 
-[mypy-bokeh.util.session_id]
+[mypy-bokeh.util.token]
 ignore_errors = False
 
 [mypy-bokeh.util.deprecation]
