+++ tests/unit/bokeh/server/test_server__server.py
@@ -38,7 +38,7 @@ from bokeh.core.properties import List, String
 from bokeh.model import Model
 from bokeh.server.server import BaseServer, Server
 from bokeh.server.tornado import BokehTornado
-from bokeh.util.session_id import (
+from bokeh.util.token import (
     check_token_signature,
     generate_jwt_token,
     get_session_id,
