+++ tests/unit/bokeh/test_client_server.py
@@ -42,7 +42,7 @@ from bokeh.document import Document
 from bokeh.document.events import ModelChangedEvent, TitleChangedEvent
 from bokeh.model import Model
 from bokeh.models import Plot
-from bokeh.util.session_id import generate_jwt_token
+from bokeh.util.token import generate_jwt_token
 from server._util_server import http_get, url, websocket_open, ws_url
 
 # Module under test
